requirements.txt
click==8.1.7
colorama==0.4.6
digi-xbee==1.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
pypacker==5.4
pyserial==3.5
python-pytun==2.4.1
rich==13.7.0
shellingham==1.5.4
tomlkit==0.12.3
typer==0.9.0
typing_extensions==4.9.0
